+++ lib/main.dart
@@ -24,6 +24,9 @@ part 'widget/game_history.dart';
 part 'widget/bidding.dart';
 part 'widget/common.dart';
 part 'middleware/middleware.dart';
+part 'middleware/room_middleware.dart';
+part 'middleware/game_middleware.dart';
+part 'middleware/bidding_middleware.dart';
 part 'reducers/heist_reducers.dart';
 part 'reducers/player_reducers.dart';
 part 'reducers/reducers.dart';
